requirements.txt
openai
python-dotenv
PySide6